# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Visca Corporation API",
        "description": "Dokumentasi Open Api Visca Corporation",
        "contact": {
            "email": "bimasaktiputra95@gmail.com"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Server Local"
        }
    ],
    "paths": {
...
        "/v1/transaksi/list-produk": {
            "get": {
                "tags": [
                    "Transaksi"
                ],
                "summary": "List Data Produk Transaksi",
                "description": "List Data Produk Transaksi",
                "operationId": "List Data Produk Transaksi",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Halaman page data untuk ditampilkan",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    },
                    {
                        "name": "category_id",
                        "in": "query",
                        "description": "Sort data berdasarkan id ketegori",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Tipe pencarian berdasarkan nama produk",
                        "required": false,
                        "schema": {
                            "type": "text"
                        },
                        "example": "Nasi Goreng Ayam"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List Data Produk Transaksi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "string",
                                                "example": [
                                                    "List Data Produk Transaksi"
                                                ]
                                            },
                                            "collectionFormat": "multi"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/v1/transaksi/buat-transaksi": {
            "post": {
                "tags": [
                    "Transaksi"
                ],
                "summary": "Buat Transaksi Penjualan ke customer open bill / transaksi lunas",
                "description": "Buat Transaksi Penjualan ke customer open bill / transaksi lunas",
                "operationId": "Buat Transaksi Penjualan ke customer open bill / transaksi lunas",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "transaction_id",
                                    "client_id",
                                    "taxer_id",
                                    "discount_id",
                                    "discount_temp_id",
                                    "other_id",
                                    "other_temp_id",
                                    "taxrate",
                                    "biaya_layanan",
                                    "discount_nom",
                                    "biaya_tambahan",
                                    "grandtotal",
                                    "amount",
                                    "method",
                                    "tipe_pembayaran"
                                ],
                                "properties": {
                                    "transaction_id": {
                                        "description": "0 Untuk transaksi baru",
                                        "type": "integer"
                                    },
                                    "client_id": {
                                        "description": "Id client / id pembeli",
                                        "type": "integer"
                                    },
                                    "grandtotal": {
                                        "description": "Jumlah grandtotal penjualan",
                                        "type": "integer"
                                    },
                                    "amount": {
                                        "description": "Jumlah pembayaran transaksi",
                                        "type": "integer"
                                    },
                                    "payment_id": {
                                        "description": "Id Payment yang akan digunakan",
                                        "type": "integer"
                                    },
                                    "details": {
                                        "description": "Details data transaksi produk",
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "id": {
                                                    "description": "Id detail penjualan, 0 untuk menambahkan produk transaksi untuk transaksi open bill",
                                                    "type": "integer"
                                                },
                                                "product_id": {
                                                    "description": "id produk terpilih yang ingin di transaksikan",
                                                    "type": "integer"
                                                },
                                                "price": {
                                                    "description": "Harga jual produk",
                                                    "type": "integer"
                                                },
                                                "qty": {
                                                    "description": "Kuantitas Pembelian Produk",
                                                    "type": "string"
                                                },
                                                "total": {
                                                    "description": "Total harga per kuantitas produknya",
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Berhasil membuat transaksi baru",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolval",
                                            "example": "true"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Berhasil membuat transaksi baru"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Transaksi gagal, cek kembali no reff dan validasi anda",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Transaksi gagal, cek kembali no reff dan validasi anda"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}